Completing the process of setting up redux by creating a container that:
Imports Hello Component for the functionality, importing the files from actions as an interface to the data transaction process, Importing boundaries for the information rules, and importing the dispath and connect functions from redux.

the function mapStateToProps() will partially format the information to be processed by redux
the function mapDispatchToProps() allows us to create a connection for our app to the actions well need it to complete. In this situation the actions are incrementing and decrementing. When map these properties (actions) with the dispatch we will have access to it with a callback function.

the connect function brings everything to fruit by connecting to redux.